dashboard-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.0.3
boto3==1.28.45
python-dotenv==1.0.0
gunicorn==20.1.0